fix: test case

+++ sequel/types/int_test.go
@@ -18,14 +18,14 @@ func TestInteger(t *testing.T) {
 		t.Run("int16", func(t *testing.T) {
 			var i16 int16
 			v := Integer(&i16)
-			require.NoError(t, v.Scan(-68))
+			require.NoError(t, v.Scan(int64(-68)))
 			require.Equal(t, int16(-68), v.Interface())
 		})
 
 		t.Run("int32", func(t *testing.T) {
 			var i32 int32
 			v := Integer(&i32)
-			require.NoError(t, v.Scan(-128))
+			require.NoError(t, v.Scan(int64(-128)))
 			require.Equal(t, int32(-128), v.Interface())
 		})
 	})
